requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-multipart
python-jose[cryptography]
passlib[bcrypt]
langchain
langchain-openai
langchain-community
openai
httpx
faiss-cpu
pypdf
python-docx
pandas
openpyxl
xlrd
faker
python-dotenv
docx2txt
